refactor(agents): extract AuthConfigDialog from AgentActionsPanel
